+++ Microsoft.Azure.Cosmos/src/CosmosClientOptions.cs
@@ -5,6 +5,7 @@
 namespace Microsoft.Azure.Cosmos
 {
     using System;
+    using System.Collections.Generic;
     using System.Collections.ObjectModel;
     using System.Data.Common;
     using System.Linq;
@@ -101,8 +102,18 @@ namespace Microsoft.Azure.Cosmos
         /// When this property is not specified, the SDK uses the write region as the preferred region for all operations.
         /// </remarks>
         /// <seealso cref="CosmosClientBuilder.WithApplicationRegion(string)"/>
-        /// <seealso href="https://docs.microsoft.com/azure/cosmos-db/how-to-multi-master">Configure multi-master</seealso>
+        /// <seealso href="https://docs.microsoft.com/azure/cosmos-db/high-availability#high-availability-with-cosmos-db-in-the-event-of-regional-outages">High availability on regional outages</seealso>
         public string ApplicationRegion { get; set; }
+
+        /// <summary>
+        /// Gets and sets the preferred regions for geo-replicated database accounts in the Azure Cosmos DB service. 
+        /// </summary>
+        /// <remarks>
+        /// When this property is specified, the SDK will use the region list in the provided order to define the endpoint failover order.
+        /// This configuration is an alternative to <see cref="ApplicationRegion"/>, either one can be set but not both.
+        /// </remarks>
+        /// <seealso href="https://docs.microsoft.com/azure/cosmos-db/high-availability#high-availability-with-cosmos-db-in-the-event-of-regional-outages">High availability on regional outages</seealso>
+        public IReadOnlyList<string> ApplicationPreferredRegions { get; set; }
 
         /// <summary>
         /// Get or set the maximum number of concurrent connections allowed for the target
@@ -394,7 +405,7 @@ namespace Microsoft.Azure.Cosmos
         /// <remarks>
         /// When the value of this property is false, the SDK will automatically discover write and read regions, and use them when the configured application region is not available.
         /// When set to true, availability is limited to the endpoint specified on the CosmosClient constructor.
-        /// Defining the <see cref="ApplicationRegion"/> is not allowed when setting the value to true.
+        /// Defining the <see cref="ApplicationRegion"/> or <see cref="ApplicationPreferredRegions"/>  is not allowed when setting the value to true.
         /// </remarks>
         /// <seealso href="https://docs.microsoft.com/azure/cosmos-db/high-availability">High availability</seealso>
         public bool LimitToEndpoint { get; set; } = false;
@@ -403,6 +414,17 @@ namespace Microsoft.Azure.Cosmos
         /// Allows optimistic batching of requests to service. Setting this option might impact the latency of the operations. Hence this option is recommended for non-latency sensitive scenarios only.
         /// </summary>
         public bool AllowBulkExecution { get; set; }
+
+        /// <summary>
+        /// Gets or sets the flag to enable address cache refresh on TCP connection reset notification.
+        /// </summary>
+        /// <remarks>
+        /// Does not apply if <see cref="ConnectionMode.Gateway"/> is used.
+        /// </remarks>
+        /// <value>
+        /// The default value is false
+        /// </value>
+        public bool EnableTcpConnectionEndpointRediscovery { get; set; } = false;
 
         /// <summary>
         /// Gets or sets the connection protocol when connecting to the Azure Cosmos service.
@@ -533,12 +555,18 @@ namespace Microsoft.Azure.Cosmos
                 MaxRequestsPerTcpConnection = this.MaxRequestsPerTcpConnection,
                 MaxTcpConnectionsPerEndpoint = this.MaxTcpConnectionsPerEndpoint,
                 EnableEndpointDiscovery = !this.LimitToEndpoint,
-                PortReuseMode = this.portReuseMode
+                PortReuseMode = this.portReuseMode,
+                EnableTcpConnectionEndpointRediscovery = this.EnableTcpConnectionEndpointRediscovery
             };
 
             if (this.ApplicationRegion != null)
             {
                 connectionPolicy.SetCurrentLocation(this.ApplicationRegion);
+            }
+
+            if (this.ApplicationPreferredRegions != null)
+            {
+                connectionPolicy.SetPreferredLocations(this.ApplicationPreferredRegions);
             }
 
             if (this.MaxRetryAttemptsOnRateLimitedRequests != null)
@@ -637,6 +665,16 @@ namespace Microsoft.Azure.Cosmos
             if (!string.IsNullOrEmpty(this.ApplicationRegion) && this.LimitToEndpoint)
             {
                 throw new ArgumentException($"Cannot specify {nameof(this.ApplicationRegion)} and enable {nameof(this.LimitToEndpoint)}. Only one can be set.");
+            }
+
+            if (this.ApplicationPreferredRegions?.Count > 0 && this.LimitToEndpoint)
+            {
+                throw new ArgumentException($"Cannot specify {nameof(this.ApplicationPreferredRegions)} and enable {nameof(this.LimitToEndpoint)}. Only one can be set.");
+            }
+
+            if (!string.IsNullOrEmpty(this.ApplicationRegion) && this.ApplicationPreferredRegions?.Count > 0)
+            {
+                throw new ArgumentException($"Cannot specify {nameof(this.ApplicationPreferredRegions)} and {nameof(this.ApplicationRegion)}. Only one can be set.");
             }
         }