+++ Microsoft.Azure.Cosmos/src/Handler/ResponseMessage.cs
@@ -8,6 +8,7 @@ namespace Microsoft.Azure.Cosmos
     using System.Diagnostics;
     using System.IO;
     using System.Net;
+    using System.Text;
     using Microsoft.Azure.Cosmos.Resource.CosmosExceptions;
     using Microsoft.Azure.Documents;
 
@@ -99,7 +100,7 @@ namespace Microsoft.Azure.Cosmos
         /// <summary>
         /// Gets the reason for a failure in the current response.
         /// </summary>
-        public virtual string ErrorMessage => this.CosmosException?.ToString(includeDiagnostics: false);
+        public virtual string ErrorMessage => this.CosmosException?.Message;
 
         /// <summary>
         /// Gets the current <see cref="ResponseMessage"/> HTTP headers.