+++ Microsoft.Azure.Cosmos/src/Query/Core/ExecutionContext/PipelinedDocumentQueryExecutionContext.cs
@@ -9,7 +9,6 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionContext
     using System.Threading.Tasks;
     using Microsoft.Azure.Cosmos;
     using Microsoft.Azure.Cosmos.CosmosElements;
-    using Microsoft.Azure.Cosmos.Json;
     using Microsoft.Azure.Cosmos.Query.Core.ContinuationTokens;
     using Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent;
     using Microsoft.Azure.Cosmos.Query.Core.ExecutionComponent.Aggregate;
@@ -84,6 +83,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionContext
     /// This bubbles down until you reach a component that has a DocumentProducer that fetches a document from the backend.
     /// </para>
     /// </summary>
+#nullable enable
     internal sealed class PipelinedDocumentQueryExecutionContext : CosmosQueryExecutionContext
     {
         /// <summary>
@@ -225,7 +225,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionContext
                 };
             }
 
-            if (queryInfo.HasOffset)
+            if (queryInfo.Offset.HasValue)
             {
                 Func<CosmosElement, Task<TryCatch<IDocumentQueryExecutionComponent>>> tryCreateSourceAsync = tryCreatePipelineAsync;
                 tryCreatePipelineAsync = async (continuationToken) =>
@@ -238,7 +238,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionContext
                 };
             }
 
-            if (queryInfo.HasLimit)
+            if (queryInfo.Limit.HasValue)
             {
                 Func<CosmosElement, Task<TryCatch<IDocumentQueryExecutionComponent>>> tryCreateSourceAsync = tryCreatePipelineAsync;
                 tryCreatePipelineAsync = async (continuationToken) =>
@@ -251,7 +251,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionContext
                 };
             }
 
-            if (queryInfo.HasTop)
+            if (queryInfo.Top.HasValue)
             {
                 Func<CosmosElement, Task<TryCatch<IDocumentQueryExecutionComponent>>> tryCreateSourceAsync = tryCreatePipelineAsync;
                 tryCreatePipelineAsync = async (continuationToken) =>
@@ -311,7 +311,7 @@ namespace Microsoft.Azure.Cosmos.Query.Core.ExecutionContext
                     return queryResponse;
                 }
 
-                string updatedContinuationToken;
+                string? updatedContinuationToken;
                 if (queryResponse.DisallowContinuationTokenMessage == null)
                 {
                     if (queryResponse.ContinuationToken != null)