+++ Microsoft.Azure.Cosmos/src/Resource/ClientContextCore.cs
@@ -5,6 +5,7 @@
 namespace Microsoft.Azure.Cosmos
 {
     using System;
+    using System.Diagnostics;
     using System.IO;
     using System.Net.Http;
     using System.Text;
@@ -323,6 +324,59 @@ namespace Microsoft.Azure.Cosmos
             return this.batchExecutorCache.GetExecutorForContainer(container, this);
         }
 
+        internal override async Task<Stream> EncryptItemAsync(
+            Stream input,
+            EncryptionOptions encryptionOptions,
+            DatabaseCore database,
+            CosmosDiagnosticsContext diagnosticsContext,
+            CancellationToken cancellationToken)
+        {
+            if (input == null)
+            {
+                throw new ArgumentException(ClientResources.InvalidRequestWithEncryptionOptions);
+            }
+
+            Debug.Assert(encryptionOptions != null);
+            Debug.Assert(database != null);
+            Debug.Assert(diagnosticsContext != null);
+
+            using (diagnosticsContext.CreateScope("Encrypt"))
+            {
+                return await this.EncryptionProcessor.EncryptAsync(
+                    input,
+                    encryptionOptions,
+                    database,
+                    this.ClientOptions.EncryptionKeyWrapProvider,
+                    diagnosticsContext,
+                    cancellationToken);
+            }
+        }
+
+        internal override async Task<Stream> DecryptItemAsync(
+            Stream input,
+            DatabaseCore database,
+            CosmosDiagnosticsContext diagnosticsContext,
+            CancellationToken cancellationToken)
+        {
+            if (input == null || this.ClientOptions.EncryptionKeyWrapProvider == null)
+            {
+                return input;
+            }
+
+            Debug.Assert(database != null);
+            Debug.Assert(diagnosticsContext != null);
+
+            using (diagnosticsContext.CreateScope("Decrypt"))
+            {
+                return await this.EncryptionProcessor.DecryptAsync(
+                    input,
+                    database,
+                    this.ClientOptions.EncryptionKeyWrapProvider,
+                    diagnosticsContext,
+                    cancellationToken);
+            }
+        }
+
         public override void Dispose()
         {
             this.Dispose(true);
@@ -394,7 +448,7 @@ namespace Microsoft.Azure.Cosmos
 
         private static HttpClientHandler CreateHttpClientHandler(CosmosClientOptions clientOptions)
         {
-            if (clientOptions == null || (clientOptions.WebProxy == null))
+            if (clientOptions == null || clientOptions.WebProxy == null)
             {
                 return null;
             }